agents_orchestrator/requirements.txt
crewai>=0.70.0
crewai-tools>=0.12.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
google-generativeai>=0.8.0